backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asgiref==3.9.1",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.3",
    "cryptography==45.0.6",
    "dj-database-url==3.0.1",
    "django==5.2.5",
    "django-cors-headers==4.7.0",
    "django-extensions==4.1",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "et-xmlfile==2.0.0",
    "gunicorn==23.0.0",
    "idna==3.10",
    "openpyxl==3.1.5",
    "packaging==25.0",
    "pdfminer-six==20250506",
    "pdfplumber==0.11.7",
    "pillow==11.3.0",
    "psycopg2-binary==2.9.10",
    "pycparser==2.22",
    "pyjwt==2.10.1",
    "pypdfium2==4.30.0",
    "python-decouple==3.8",
    "python-dotenv==1.1.1",
    "requests==2.32.5",
    "sqlparse==0.5.3",
    "stripe==12.4.0",
    "typing-extensions==4.15.0",
    "tzdata==2025.2",
    "urllib3==2.5.0",
]
